pyproject.toml
[project]
name = "protein-lang"
version = "0.6.1"
description = "Protein: a macro language for data composition and templating"
readme = "README.md"
authors = [{ name = "Laurent" }]
requires-python = ">=3.11"
dependencies = [
    "hjson>=3.1.0",
    "jinja2>=3.1.6",
    "jsonschema>=4.25.1",
    "keyring>=25.7.0",
    "markdown-it-py>=4.0.0",
    "mdit-py-plugins>=0.5.0",
    "python-frontmatter>=1.1.0",
    "rich>=14.2.0",
    "ruamel-yaml>=0.18.16",
    "sqlalchemy>=2.0.45",
    "super-collections>=0.6.2",
    "tomlkit>=0.13.3",
]

license = { file = "LICENSE" }

[project.scripts]
protein = "protein.cli:main"
yamlpp = "protein.cli:main"  # temporary alias

[tool.setuptools.packages.find]
where = ["."]

[tool.uv]
package = true

[dependency-groups]
dev = ["pytest>=9.0.1", "toml-cli>=0.5.0"]


[tool.setuptools.package-data]
protein = ["protein_schema.yaml"]
test = ["source/*"]
